+++ CHANGELOG.md
@@ -5,7 +5,9 @@ All notable changes to OpenVM will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project follows a versioning principles documented in [VERSIONING.md](./VERSIONING.md).
 
-## v1.4.1 (Unreleased)
+## Unreleased
+
+## v1.4.1 (2025-10-26)
 
 ### Added
 - (CLI) Add configurable segmentation arguments to `cargo openvm prove` commands.